Update to webpack 2
This commit only takes the minimal actions that are required to update to webpack 2. The tests (and probably the loader itself) require more refactoring, but there's currently little time for that.

Also removed the synchronous compilation since it has been deprecated for a while: https://github.com/webpack-contrib/less-loader/issues/84