feat: allow user to choose between resolvers
If a paths property is set in the options, it will effectively disable webpack's resolver. This allows you to use Less' default resolver which is usually faster. However, if you choose to set the paths, you won't be able to use webpack features like resolve.alias or importing other file types.

This is not a breaking change because setting the paths hasn't worked properly before this anyway.