fix: resolve alias
Less automatically adds a .less file extension if no extension was given. This is problematic if there is a module request like @import "~some-module"; because in this case Less will call our file manager with `~some-module.less`. Since dots in module names are highly discouraged, we can safely assume that this is an error and we need to remove the .less extension again.

This means that aliases configured in the `webpack.config.js` are now resolved correctly if the webpack resolver is used.

BREAKING CHANGE: If you've already configured your `resolve.alias` with a `.less` extension, you can now remove that wrong extension.